% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_fn.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map State}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Integer}

\item{year}{Integer}
}
\value{
This function returns an object for graphical visualization for accidents' locations.
}
\description{
Plot a map displaying all accidents previously recorded in the provided state number and year.
}
\examples{
\dontrun{
fars_map_state(16, 2020)
fars_map_state(70, 2014))
}

}
