% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_fn.R
\name{fars_read}
\alias{fars_read}
\title{Read data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string of characters}
}
\value{
An R dataframe table
}
\description{
This function reads comma separated value formatted dataset and returns it as an R dataframe table.
}
\details{
This function throws an error message if the filename does not exist.
}
\examples{
\dontrun{fars_read("filename.csv")}

}
