% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_fn.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Create a months and years dataframe for a given number of years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Integers vector}
}
\value{
This function returns an aggregated dataframe of months for each year from the read files.
}
\description{
This function reads multiple comma separated values files for a given number of years, then create an aggregated view of months and years from the read files.
}
\details{
This function throws an error message if the file matching the given year does not exist.
}
\examples{
\dontrun{
fars_read_years(2020)
fars_read_years(c(2001, 2002))
}

}
