% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_fn.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize Years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Integers vector}
}
\value{
This function returns a dataframe summarizing the monthly number of accidents for each month in the given number of years.
}
\description{
This function generates a summary of the monthly number of accidents for each month in the given number of years.
}
\examples{
\dontrun{
fars_summarize_years(2020)
fars_summarize_years(c(2001, 2002))
}

}
