% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_fn.R
\name{make_filename}
\alias{make_filename}
\title{Print a given year's filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Integer}
}
\value{
This function returns a string of characters resulted from the merging of a prespecified text and a year given as parameter
}
\description{
This function prints a string of characters resulted from the merging of a prespecified text and a year given as parameter
}
\examples{
\dontrun{
make_filename(2001)
make_filename(2020)
}
}
